+++ lumen/models/group_member.py
@@ -8,8 +8,8 @@ class GroupMember(db.Model):
     )
 
     id = db.Column(db.Integer, primary_key=True)
-    group_id = db.Column(db.Integer, db.ForeignKey("groups.id"), nullable=False)
-    entity_id = db.Column(db.Integer, db.ForeignKey("entities.id"), nullable=False)
+    group_id = db.Column(db.Integer, db.ForeignKey("groups.id", ondelete="CASCADE"), nullable=False)
+    entity_id = db.Column(db.Integer, db.ForeignKey("entities.id", ondelete="CASCADE"), nullable=False)
     config_managed = db.Column(db.Boolean, default=False, nullable=False)
 
     entity = db.relationship("Entity", backref="group_memberships")