+++ lumen/models/model_config.py
@@ -12,6 +12,6 @@ class ModelConfig(db.Model):
     active = db.Column(db.Boolean, default=True, nullable=False)
     created_at = db.Column(db.DateTime, default=datetime.utcnow)
 
-    endpoints = db.relationship("ModelEndpoint", backref="model_config", lazy="dynamic")
-    limits = db.relationship("EntityModelLimit", backref="model_config", lazy="dynamic")
-    stats = db.relationship("ModelStat", backref="model_config", lazy="dynamic")
+    endpoints = db.relationship("ModelEndpoint", backref="model_config", lazy="dynamic", cascade="all, delete-orphan", passive_deletes=True)
+    limits = db.relationship("EntityModelLimit", backref="model_config", lazy="dynamic", passive_deletes=True)
+    stats = db.relationship("ModelStat", backref="model_config", lazy="dynamic", passive_deletes=True)